+++ rocketgram/connectors/aiohttpconnector.py
@@ -10,7 +10,7 @@ from json import JSONDecodeError
 import aiohttp
 
 from .connector import Connector, HEADERS
-from ..api import API_URL, Request, Response
+from ..api import API_URL, API_FILE_URL, Request, Response
 from ..errors import RocketgramNetworkError, RocketgramParseError
 
 try:
@@ -22,9 +22,10 @@ logger = logging.getLogger('rocketgram.connectors.aiohttp')
 
 
 class AioHttpConnector(Connector):
-    __slots__ = ('_api_url', '_session', '_timeout')
+    __slots__ = ('_api_url', '_api_file_url', '_session', '_timeout')
 
-    def __init__(self, *, timeout: int = 35, api_url: str = API_URL):
+    def __init__(self, *, timeout: int = 35, api_url: str = API_URL, api_file_url: str = API_FILE_URL):
+        self._api_file_url = api_file_url
         self._api_url = api_url
         self._session = aiohttp.ClientSession(loop=asyncio.get_event_loop())
         self._timeout = timeout
@@ -66,3 +67,6 @@ class AioHttpConnector(Connector):
             raise
         except Exception as error:
             raise RocketgramNetworkError(error) from error
+
+    def resolve_file_url(self, token: str, file_path: str) -> str:
+        return self._api_file_url % (token, file_path)